Add build options and Github action to check if the project builds (#5)
This PR introduces an action that checks if the project builds. If we
all agree, I would put this action as a requirement to merge PRs into
`main` and `develop`.

Examples:
- I made a
[commit](https://github.com/VDP-noclip/github-playground/actions/runs/3260541219)
that breaks the project on purpose and the [action
fails](https://github.com/VDP-noclip/github-playground/actions/runs/3260541219).
- Then, I made a commit to revert that change and opened this PR, and
the [action indeed
passes](https://github.com/VDP-noclip/github-playground/actions/runs/3260610738).

Remarks:
- It's the first time I try this action, therefore I am not sure I had
made the setup correctly. We will discover it only with time ;)
- It takes 5 minutes approx. to make the check so we might finish our
free Github actions minutes. In that case, we could add this check only
for the `main` branch.

Co-authored-by: Nicola Ischia <nicola.ischia@schindler.com>